Move shared code to subdirectory
